Share the Fork.RI.exe path to WSL and replace the Fork.RI.exe path in the argument with the path to the Fork.RI script.
The path to Fork.RI.exe is shared using an environment variable named FORK_RI_EXE_PATH that is created temporarily and passed to WSL using WSLENV.
The path to Fork.RI.exe in the argument is replaced with the path to the Fork.RI script that is located in the same directory as the wslgit executable.
The Fork.RI script then use the FORK_RI_EXE_PATH for calling the real Fork.RI.exe.